// libs/api/src/messages/message-code.enum.ts
export enum MessageCode {
  SUCCESS = 'SUCCESS',
  CREATED = 'CREATED',
  UPDATED = 'UPDATED',
  DELETED = 'DELETED',
  FETCHED = 'FETCHED',
  LIST_FETCHED = 'LIST_FETCHED',
  NOT_FOUND = 'NOT_FOUND',
  ALREADY_EXISTS = 'ALREADY_EXISTS',
  VALIDATION_FAILED = 'VALIDATION_FAILED',
  FIELD_REQUIRED = 'FIELD_REQUIRED',
  INVALID_INPUT = 'INVALID_INPUT',
  UNAUTHORIZED = 'UNAUTHORIZED',
  FORBIDDEN = 'FORBIDDEN',
  INTERNAL_ERROR = 'INTERNAL_ERROR',
}